requirements.txt
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.0
jinja2>=3.1.4
pyyaml>=6.0.1
hypothesis>=6.100.0
pytest>=8.0.0
